[+] Добавление функции скидки для заказа

+++ orders/models.py
@@ -7,7 +7,7 @@ class Discount(models.Model):
     """Модель для скидок"""
 
     name = models.CharField(max_length=100, verbose_name="Название скидки")
-    discount_percent = models.PositiveSmallIntegerField(verbose_name="Процент скидки")
+    discount_percent = models.DecimalField(max_digits=5, decimal_places=2, verbose_name="Процент скидки")
 
     def __str__(self):
         return self.name
